pyproject.toml
[project]
name = "smart-traffic-intersection"
version = "1.0.0"
description = "AI-driven 4-way traffic signal controller"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.9.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "openenv-core>=0.2.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
